feat: implemented librarian authentication

+++ src/main/java/com/johnny/libmgtbackend/service/LibrarianService.java
@@ -1,8 +1,10 @@
 package com.johnny.libmgtbackend.service;
 
+import com.johnny.libmgtbackend.auth.AuthenticationProvider;
 import com.johnny.libmgtbackend.dtos.BorrowRecordDto;
 import com.johnny.libmgtbackend.dtos.LibrarianDto;
 import com.johnny.libmgtbackend.exception.ModelNotFoundException;
+import com.johnny.libmgtbackend.exception.UnauthorizedException;
 import com.johnny.libmgtbackend.models.Book;
 import com.johnny.libmgtbackend.models.BorrowRecord;
 import com.johnny.libmgtbackend.models.Librarian;
@@ -11,7 +13,6 @@ import com.johnny.libmgtbackend.repository.BookRepository;
 import com.johnny.libmgtbackend.repository.BorrowRepository;
 import com.johnny.libmgtbackend.repository.LibrarianRepository;
 import com.johnny.libmgtbackend.repository.PatronRepository;
-import com.johnny.libmgtbackend.request.CreateLibrarianRequest;
 import com.johnny.libmgtbackend.request.UpdateLibrarianRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -36,6 +37,9 @@ public class LibrarianService {
     private BorrowRepository borrowRepository;
 
     @Autowired
+    private AuthenticationProvider authenticationProvider;
+
+    @Autowired
     private PagedResourcesAssembler<LibrarianDto> pagedResourcesAssembler;
 
     public PagedModel<EntityModel<LibrarianDto>> getAllLibrarians(Pageable paging) {
@@ -48,54 +52,68 @@ public class LibrarianService {
         return new LibrarianDto(librarianRepository.findById(id).orElseThrow(() -> new ModelNotFoundException(Librarian.class, id)));
     }
 
-    public LibrarianDto createLibrarian(CreateLibrarianRequest request) {
-        var librarian = new Librarian(request.email, request.name, request.password);
+    public LibrarianDto updateLibrarian(UpdateLibrarianRequest request, Long id) throws Exception {
+        var authLibrarian = authenticationProvider.getAuthenticatedLibrarian();
+        var librarian = librarianRepository.findById(id).orElseThrow(() -> new ModelNotFoundException(Librarian.class, id));
+
+        if (!authLibrarian.getId().equals(librarian.getId())) {
+            throw new UnauthorizedException("You cannot update another librarian's profile!");
+        }
+
+        if (request.email != null && !librarian.getEmail().equals(request.email)) {
+            var existingLibrarian = librarianRepository.findByEmail(request.email);
+            if (existingLibrarian != null) {
+                throw new Exception("Librarian with email already exists");
+            }
+            librarian.setEmail(request.email);
+        }
+        if (request.name != null && !librarian.getName().equals(request.name)) {
+            librarian.setName(request.name);
+        }
+
         return new LibrarianDto(librarianRepository.save(librarian));
     }
 
-    public LibrarianDto updateLibrarian(UpdateLibrarianRequest request, Long id) {
-        var user = librarianRepository.findById(id).orElseThrow(() -> new ModelNotFoundException(Librarian.class, id));
-
-        if (request.email != null && !user.getEmail().equals(request.email)) {
-            user.setEmail(request.email);
+    public void deleteLibrarian(Long id) throws Exception {
+        var authLibrarian = authenticationProvider.getAuthenticatedLibrarian();
+        if (!authLibrarian.getId().equals(id)) {
+            throw new UnauthorizedException("You cannot delete another person's profile");
         }
-        if (request.name != null && !user.getName().equals(request.name)) {
-            user.setName(request.name);
-        }
-        if (request.password != null && !user.getPassword().equals(request.password)) {
-            user.setPassword(request.password);
-        }
-
-        return new LibrarianDto(librarianRepository.save(user));
-    }
-
-    public void deleteLibrarian(Long id) {
         librarianRepository.deleteById(id);
     }
 
-    public BorrowRecordDto borrowBook(Long bookId, Long patronId) {
+    public BorrowRecordDto borrowBook(Long bookId, Long patronId) throws Exception {
+        var authLibrarian = authenticationProvider.getAuthenticatedLibrarian();
+
         var book = bookRepository.findById(bookId).orElseThrow(() -> new ModelNotFoundException(Book.class, bookId));
         var patron = patronRepository.findById(patronId).orElseThrow(() -> new ModelNotFoundException(Patron.class, patronId));
 
         var borrowRecord = borrowRepository.findBorrowRecordByBookAndPatronAndReturnDateNull(book, patron);
+
         if (borrowRecord == null) {
-            var newRecord = new BorrowRecord(book, patron);
+            var newRecord = new BorrowRecord(book, patron, authLibrarian);
             newRecord.setBorrowDate(LocalDate.now());
 
             return new BorrowRecordDto(borrowRepository.save(newRecord));
         }
-        return null;
+        throw new UnauthorizedException("You cannot borrow the same book twice");
     }
 
-    public BorrowRecordDto returnBook(Long bookId, Long patronId) {
+    public BorrowRecordDto returnBook(Long bookId, Long patronId) throws Exception {
+        var authLibrarian = authenticationProvider.getAuthenticatedLibrarian();
+
         var book = bookRepository.findById(bookId).orElseThrow(() -> new ModelNotFoundException(Book.class, bookId));
         var patron = patronRepository.findById(patronId).orElseThrow(() -> new ModelNotFoundException(Patron.class, patronId));
 
         var borrowRecord = borrowRepository.findBorrowRecordByBookAndPatronAndReturnDateNull(book, patron);
+
         if (borrowRecord != null) {
+            if (!borrowRecord.getLibrarian().getId().equals(authLibrarian.getId())) {
+                throw new UnauthorizedException("Kindly return the book to the Librarian that borrowed it to you");
+            }
             borrowRecord.setReturnDate(LocalDate.now());
             return new BorrowRecordDto(borrowRepository.save(borrowRecord));
         }
-        return null;
+        throw  new ModelNotFoundException(BorrowRecord.class, book.getTitle(), patron.getName());
     }
 }
